% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cw5.R
\name{collatz}
\alias{collatz}
\title{Collatz}
\usage{
collatz(c0, print = FALSE)
}
\arguments{
\item{print}{\cr
Decyduje czy wyswietlac informacje o dlugosci i maksymalnej wartosci ciagu}

\item{c}{\cr
Zadany parametr}
}
\value{
Lista zawierające wszystkie elementy ciągu przed wpadnięciem w cykl
}
\description{
Funkcja \code{collatz} zwraca wszystkie elementy ciągu przed wpadnięciem w cykl
}
\examples{
collatz(5)
collatz(9)

}
\author{
Szymon Kubica 264068
}
