% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cw4.R
\name{fibonacci}
\alias{fibonacci}
\title{Fibonacci}
\usage{
fibonacci(n, iter = FALSE)
}
\arguments{
\item{n}{\cr
Liczba okreslajaca ile wyrazow ciagu fibonnaciego zwrocic}

\item{iter}{\cr
Gdy ustawione na TRUE, uzywana jest wersja iteracyjna}
}
\value{
Lista zawierające \code{n} pierwszych wyrazow ciagu fibonnaciego
}
\description{
Funkcja \code{fibonacci} zwraca listę \code{n} pierwszych wyrazow ciągu fibonnaciego.
Domyślnie używa rekurencji ale ustawiając parametr \code{iter} na TRUE używa wersji iteracyjnej
}
\examples{
fibonacci(5)
fibonacci(9, iter = TRUE)

}
\author{
Szymon Kubica 264068
}
