% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cw1.R
\name{heron}
\alias{heron}
\title{Oblicza pola trójkąta według wzoru Herona}
\usage{
heron(a, b, c)
}
\arguments{
\item{a}{Długość boku a (liczba dodatnia)}

\item{b}{Długość boku b (liczba dodatnia)}

\item{c}{Długość boku c (liczba dodatnia)}
}
\value{
Pole trójkąta
}
\description{
Funkcja \code{heron} oblicza pole trójkąta dla podanych długości boków
za pomocą wozru Herona. Sprawdza poprawność argumentów i
istnienie trójkąta.
}
\examples{
heron(3,4,5) # = 6
heron(5,5,6) # =~12

}
\author{
Szymon Kubica 264068
}
