% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cw6.R
\name{komplement}
\alias{komplement}
\title{Komplement}
\usage{
komplement(dna)
}
\arguments{
\item{dna}{\cr
Ciąg znaków DNA}
}
\value{
Nić matrycowa
}
\description{
Funkcja \code{komplement} zwraca nić matrycową DNA na podstawie nici kodującej
}
\examples{
komplement("TTAA")

}
\author{
Szymon Kubica 264068
}
