% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cw3.R
\name{podzbiory}
\alias{podzbiory}
\title{Podzbiory}
\usage{
podzbiory(x)
}
\arguments{
\item{x}{\cr
Zbiór z którego będą zwracane podzbiory (wektor liczb, tekstów itp.).}
}
\value{
Lista zawierające wszystkie mozliwe kombinacje podzbiorów
}
\description{
Funkcja \code{podzbiory} zwraca listę wszystkich swoich podzbiorów
}
\examples{
podzbiory(c(1,2,3)
podzbiory(c("a","b","c","d")

}
\author{
Szymon Kubica 264068
}
