% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cw6.R
\name{transkrybuj}
\alias{transkrybuj}
\title{Transkrybcja}
\usage{
transkrybuj(template_dna)
}
\arguments{
\item{dna}{\cr
Nić sekwencyjna}
}
\value{
nić RNA
}
\description{
Funkcja \code{transkrybuj} zwraca nić matrycową RNA na podstawie nici sekwencyjnej
}
\examples{
transkrybuj("ACGT")

}
\author{
Szymon Kubica 264068
}
