% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cw6.R
\name{transluj}
\alias{transluj}
\title{Transluj}
\usage{
transluj(rna_seq, code3 = TRUE)
}
\arguments{
\item{dna}{\cr
mRNA}
}
\value{
sekwencja aminokwasów
}
\description{
Funkcja \code{transkrybuj} tłumaczy mRNA na sekwencję aminokwasów
}
\examples{
transluj("AUGAAAUAA")

}
\author{
Szymon Kubica 264068
}
