% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cw2.R
\name{wspolne}
\alias{wspolne}
\title{Część wspólna dwóch multizbiorów}
\usage{
wspolne(x, y)
}
\arguments{
\item{x}{\cr
Pierwszy multizbiór (wektor liczb, tekstów itp.).}

\item{y}{\cr
Drugi multizbiór (wektor tego samego typu co \code{x})}
}
\value{
Ten sam typ co \code{x,y}: wektor zawierający elementy
wspólne, powtórzone minimalną liczbę razy. Jeżeli nie ma wspólnych
elementów, zwraca wektor długośći 0.
}
\description{
Funkcja \code{wspolne} zwraca multizbiór bedący częścią wspólną
dwóch podanych wektorów \code{x} i \code{y}.
}
\examples{
wspolne(c(1,2,3), c(2,2,4)) # -> c(2)
wspolne(c("a","b","a"), c("b","b")) # -> "b"

}
\author{
Szymon Kubica 264068
}
